+++ src/Ema/Generate.hs
@@ -19,10 +19,11 @@ import Ema.Asset (Asset (..))
 import Ema.CLI (crash)
 import Ema.Route.Class (IsRoute (RouteModel, allRoutes, routeEncoder))
 import Ema.Route.Encoder (
+  applyRouteEncoder,
   checkRouteEncoderGivenRoute,
-  encodeRoute,
  )
 import Ema.Site (EmaSite (siteOutput))
+import Optics.Core (review)
 import System.Directory (copyFile, createDirectoryIfMissing, doesDirectoryExist, doesFileExist, doesPathExist)
 import System.FilePath (takeDirectory, (</>))
 import System.FilePattern.Directory (getDirectoryFiles)
@@ -67,6 +68,7 @@ generateSiteFromModel' ::
   m [FilePath]
 generateSiteFromModel' dest model = do
   let enc = routeEncoder @r
+      rp = applyRouteEncoder enc model
   -- Sanity checks
   unlessM (liftIO $ doesDirectoryExist dest) $ do
     throwError $ "Destination directory does not exist: " <> toText dest
@@ -81,8 +83,8 @@ generateSiteFromModel' dest model = do
   -- Enumerate and write all routes.
   log LevelInfo $ "Writing " <> show (length routes) <> " routes"
   fmap concat . forM routes $ \r -> do
-    let fp = dest </> encodeRoute enc model r
-    case siteOutput enc model r of
+    let fp = dest </> review rp r
+    case siteOutput rp model r of
       AssetStatic staticPath -> do
         liftIO (doesPathExist staticPath) >>= \case
           True ->